refactor(runtime): extract tool_resolution module from agent_loop
`LAZY_TOOLS_THRESHOLD`, `resolve_request_tools`, and the
`ResolvedToolsCache` struct (with its `new` / `get` methods) move into
`agent_loop/tool_resolution.rs`. All three become `pub(super)` so the
main loop in `mod.rs` can keep its existing call shape
(`ResolvedToolsCache::new(...)` at the per-turn init site,
`tools_cache.get(...)` per iteration).

`resolve_request_tools` and `LAZY_TOOLS_THRESHOLD` are only consumed
from the in-file `#[cfg(test)] mod tests` block, so they are imported
into the test module directly (`use super::tool_resolution::{...};`)
rather than at file scope. This keeps `unused_imports` quiet on the
release lane — the lint can't see through `use super::*;` to a child
test module — without introducing a `#[allow]`.

Verbatim function bodies; only visibility changes. The lazy-tool
gating contract from #3044 (require `tool_load` in the granted set
before stripping non-native tools) and the per-iteration `Arc` cache
from #3586 are preserved exactly. mod.rs: 13,623 -> 13,531 LOC (-92).
cargo check + cargo clippy --lib both clean.

Refs #3710
